[C2RKMpiDec]: Add support error frame drop output
In APP ClassIn playback module, lost frame occurs in the input stream
since server instability. In this case, It is hope that error frame
be dropped and not displayed, so we add FLAG_DROP_FRAME to output work
and enable MPP error frame mode.

Type: Fix
Redmine ID: #451086
Associated modifications: N/A
Test: self-testing pass
Version: 1.14.1_[5]

Signed-off-by: Chen Jinsen <kevin.chen@rock-chips.com>
Change-Id: I4e8d44cbb67aede9ce4d233bad1871a85bce4584